refactor: add login and signup schemas into separate files
