+++ src/features/auth/loginForm/LoginForm.tsx
@@ -3,17 +3,11 @@ import { useForm } from 'react-hook-form'
 import { Button } from '@/common/components/button'
 import { ControlledCheckbox } from '@/common/components/controlled'
 import { ControlledInput } from '@/common/components/controlled/controlledInput'
-import { emailSchema, passwordSchema, rememberMeSchema } from '@/common/utils'
+import { loginSchema } from '@/features/auth/utils/schemas'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { z } from 'zod'
 
 import s from './loginForm.module.scss'
-
-const loginSchema = z.object({
-  email: emailSchema,
-  password: passwordSchema,
-  rememberMe: rememberMeSchema,
-})
 
 type Props = {
   onSubmit: (data: FormData) => void
