+++ src/features/auth/signUp/SignUpForm.tsx
@@ -6,25 +6,13 @@ import { Card } from '@/common/components/card'
 import { ControlledTextField } from '@/common/components/controlled'
 import { Typography } from '@/common/components/typography'
 import { path } from '@/common/enums'
-import { confirmPasswordSchema, emailSchema, passwordSchema, passwordsMatch } from '@/common/utils'
+import { signupSchema } from '@/features/auth/utils/schemas'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { z } from 'zod'
 
 import s from './signUpForm.module.scss'
 
-const loginScheme = z
-  .object({
-    confirmPassword: confirmPasswordSchema,
-    email: emailSchema,
-    password: passwordSchema,
-  })
-  .superRefine((data, ctx) => {
-    const issues = passwordsMatch(data)
-
-    issues.forEach(issue => ctx.addIssue(issue))
-  })
-
-export type SignUp = z.infer<typeof loginScheme>
+export type SignUp = z.infer<typeof signupSchema>
 
 type Props = {
   onSubmit: (data: SignUp) => void
@@ -32,7 +20,7 @@ type Props = {
 
 export const SignUpForm = ({ onSubmit }: Props) => {
   const { control, handleSubmit } = useForm<SignUp>({
-    resolver: zodResolver(loginScheme),
+    resolver: zodResolver(signupSchema),
   })
 
   const formSubmitHandler = handleSubmit(onSubmit)
