+++ src/features/auth/utils/schemas/index.ts
@@ -1,0 +1,2 @@
+export * from './loginSchema'
+export * from './signupSchema'
