+++ src/features/auth/utils/schemas/loginSchema.ts
@@ -1,0 +1,8 @@
+import { emailSchema, passwordSchema, rememberMeSchema } from '@/common/utils'
+import { z } from 'zod'
+
+export const loginSchema = z.object({
+  email: emailSchema,
+  password: passwordSchema,
+  rememberMe: rememberMeSchema,
+})
